refactor everything redux under finder instead of files
