+++ src/redux/actions/finder-actions.ts
@@ -1,5 +1,5 @@
-const GET_FILE_LIST = "opacity/files/get-file-list";
-const SET_LIST = "opacity/files/set-list";
+const GET_FILE_LIST = "opacity/finder/get-file-list";
+const SET_LIST = "opacity/finder/set-list";
 
 const ACTIONS = Object.freeze({
   GET_FILE_LIST,
