+++ src/redux/epics/finder-epic.ts
@@ -2,7 +2,7 @@ import { from, of } from "rxjs";
 import { map, switchMap, catchError } from "rxjs/operators";
 import { ofType, combineEpics } from "redux-observable";
 
-import filesActions from "../actions/files-actions";
+import finderActions from "../actions/finder-actions";
 import authenticationActions from "../actions/authentication-actions";
 import uploadActions from "../actions/upload-actions";
 import removeActions from "../actions/remove-actions";
@@ -10,7 +10,7 @@ import removeActions from "../actions/remove-actions";
 const getFileListEpic = (action$, state$, dependencies$) =>
   action$.pipe(
     ofType(
-      filesActions.GET_FILE_LIST,
+      finderActions.GET_FILE_LIST,
       uploadActions.UPLOAD_SUCCESS,
       removeActions.REMOVE_SUCCESS
     ),
@@ -18,8 +18,8 @@ const getFileListEpic = (action$, state$, dependencies$) =>
       const { masterHandle, folder } = payload;
 
       return from(masterHandle.getFolderMeta(folder)).pipe(
-        map((data: any) => filesActions.setList({ list: data.files })),
-        catchError(() => of(filesActions.setList({ list: [] })))
+        map((data: any) => finderActions.setList({ list: data.files })),
+        catchError(() => of(finderActions.setList({ list: [] })))
       );
     })
   );
@@ -27,7 +27,7 @@ const getFileListEpic = (action$, state$, dependencies$) =>
 const getAccountDataEpic = (action$, state$, dependencies$) =>
   action$.pipe(
     ofType(
-      filesActions.GET_FILE_LIST,
+      finderActions.GET_FILE_LIST,
       uploadActions.UPLOAD_SUCCESS,
       removeActions.REMOVE_SUCCESS
     ),
