+++ src/redux/epics/index.ts
@@ -3,7 +3,7 @@ import { combineEpics } from "redux-observable";
 import coinMarketCapEpic from "./coinmarketcap-epic";
 import metamaskEpic from "./metamask-epic";
 import signupEpic from "./signup-epic";
-import filesEpic from "./files-epic";
+import finderEpic from "./finder-epic";
 import authenticationEpic from "./authentication-epic";
 import uploadEpic from "./upload-epic";
 import downloadEpic from "./download-epic";
@@ -14,7 +14,7 @@ export default combineEpics(
   coinMarketCapEpic,
   metamaskEpic,
   signupEpic,
-  filesEpic,
+  finderEpic,
   authenticationEpic,
   uploadEpic,
   downloadEpic,
