+++ src/redux/reducers/index.ts
@@ -3,7 +3,7 @@ import { connectRouter } from "connected-react-router";
 
 import coinMarketCap from "./coinmarketcap-reducer";
 import signup from "./signup-reducer";
-import files from "./files-reducer";
+import finder from "./finder-reducer";
 import authentication from "./authentication-reducer";
 
 export default history =>
@@ -11,6 +11,6 @@ export default history =>
     router: connectRouter(history),
     coinMarketCap,
     signup,
-    files,
+    finder,
     authentication
   });
